fix sso dialog bug
