+++ app/src/main/scala/com/waz/zclient/appentry/FragmentTransactionHelper.scala
@@ -1,0 +1,24 @@
+package com.waz.zclient.appentry
+
+import androidx.fragment.app.{Fragment, FragmentActivity, FragmentTransaction}
+import com.waz.zclient.R
+
+class FragmentTransactionHelper() {
+  def showFragment(activity: FragmentActivity, f: => Fragment, tag: String, animated: Boolean = true, layoutId: Int): Unit = {
+    val transaction = activity.getSupportFragmentManager.beginTransaction()
+    if (animated) setDefaultAnimation(transaction)
+    transaction
+      .replace(layoutId, f, tag)
+      .addToBackStack(tag)
+      .commit
+  }
+
+  private def setDefaultAnimation(transaction: FragmentTransaction): FragmentTransaction = {
+    transaction.setCustomAnimations(
+      R.anim.fragment_animation_second_page_slide_in_from_right,
+      R.anim.fragment_animation_second_page_slide_out_to_left,
+      R.anim.fragment_animation_second_page_slide_in_from_left,
+      R.anim.fragment_animation_second_page_slide_out_to_right)
+    transaction
+  }
+}
