+++ app/src/main/scala/com/waz/zclient/appentry/SSOFragment.scala
@@ -22,7 +22,7 @@ import androidx.fragment.app.Fragment
 import com.waz.api.impl.ErrorResponse
 import com.waz.log.BasicLogging.LogTag.DerivedLogTag
 import com.waz.service.SSOService
-import com.waz.zclient.InputDialog.{Event, OnNegativeBtn, OnPositiveBtn, ValidatorResult}
+import com.waz.zclient.InputDialog.{Event, OnCancel, OnNegativeBtn, OnPositiveBtn, ValidatorResult}
 import com.waz.zclient._
 import com.waz.zclient.appentry.DialogErrorMessage.GenericDialogErrorMessage
 import com.waz.zclient.common.controllers.UserAccountsController
@@ -42,9 +42,11 @@ trait SSOFragment extends FragmentHelper with DerivedLogTag {
   private lazy val ssoService             = inject[SSOService]
   private lazy val userAccountsController = inject[UserAccountsController]
 
+  private def hasToken(): Future[Boolean] = userAccountsController.ssoToken.head.map(_.isDefined)
+
   private lazy val dialogStaff = new InputDialog.Listener with InputDialog.InputValidator {
     override def onDialogEvent(event: Event): Unit = event match {
-      case OnNegativeBtn          => activity.onNegativeClicked()
+      case OnNegativeBtn | OnCancel => hasToken().map(activity.onSSODialogDismissed(_))
       case OnPositiveBtn(input)   => verifyInput(input)
     }
 
@@ -56,7 +58,7 @@ trait SSOFragment extends FragmentHelper with DerivedLogTag {
   override def onStart(): Unit = {
     super.onStart()
     findChildFragment[InputDialog](SSODialogTag).foreach(_.setListener(dialogStaff).setValidator(dialogStaff))
-    extractTokenAndShowSSODialog()
+    extractTokenAndShowSSODialog(showSsoByDefault)
   }
 
   private def extractTokenFromClipboard: Future[Option[String]] = Future {
@@ -65,6 +67,8 @@ trait SSOFragment extends FragmentHelper with DerivedLogTag {
       token         <- ssoService.extractToken(clipboardText.toString)
     } yield token
   }
+
+  protected def showSsoByDefault = false
 
   protected def extractTokenAndShowSSODialog(showIfNoToken: Boolean = false): Unit =
     userAccountsController.ssoToken.head.foreach {
@@ -119,12 +123,12 @@ trait SSOFragment extends FragmentHelper with DerivedLogTag {
       }
     }
 
-  protected def activity: ActivityWithFragment = getActivity.asInstanceOf[ActivityWithFragment]
+  protected def activity: SSOFragmentHandler = getActivity.asInstanceOf[SSOFragmentHandler]
 
   protected def onVerifyingToken(verifying: Boolean): Unit = inject[SpinnerController].showSpinner(verifying)
 }
 
-trait ActivityWithFragment {
+trait SSOFragmentHandler {
   def showFragment(f: => Fragment, tag: String, animated: Boolean = true): Unit
-  def onNegativeClicked(): Unit = {}
+  def onSSODialogDismissed(hasToken: Boolean): Unit = {}
 }
